Add support for tab key to insert tabs.
The goal is to make it easier to type tab characters. Pressing the tab
key will now insert tab characters inside the text editors.

To go to next element, which was previously done with tab, can be done
with Alt+Tab (on OS X).

This change is experimental can may be reverted if it leads to a bad
user experience.