assets/script: Update upload endpoint, response scheme.
The upload API endpoint is now at /api/usercontent.

The response object contains the image URL, not Name.

Iterate over all items looking for a matching image/png file, instead
of only looking at the first item.

If available, get a plain/text string and use it as image name.
Use PreventDefault() on the event when it's being handled in order
to prevent that text from being erroneously pasted alongside with
the image.