Use context from http.Request.
In Go 1.7, each http.Request comes with a context associated with it.
Previously, an external library (such as github.com/gorilla/context)
had to be used to provide this functionality, and so one of the options
was to specify how to extract a context from an *http.Request.

Since it's a part of the standard library, it's possible for everyone
to use it, and this no longer needs to be provided externally or
customizable.

This simplifies and improves the API slightly.