Support highlighting issues with unread notifications.
If an optional notifications service is provided in issuesapp.Options,
then issuesapp will determine if any of the issues in current repository
have unread notifications for the current user. If so, those issues
will be highlighted with a thin blue marker on the left side when
viewing a list of all issues.

This makes it easier to spot which issues have new unread updates
when looking at the list of all issues (rather than at the list of
all notifications).

Regenerate.