Detect issue not found situation, serve 404.
There's currently no great standard for issues.Service implementations
to indicate an issue or comment isn't found. Normally, it should be
os.IsNotExit(err) compatible, but that's seemingly not possible with
html/template.

That shouldn't be a problem after a future refactor, but detect a
specific common case for now.

Fixes shurcooL/home#3.