Remove Options.{RepoSpec,BaseURI}; rely on context key parameters.
RepoSpec, BaseURI are now expected to be provided via request context
values, using RepoSpecContextKey and BaseURIContextKey.

Using context keys to provide required parameters to an http.Handler is
a little unusual and experimental. There may be room for future
improvement. However, this is an improvement over the previous approach
of having custom funcs that must extract this information from the
request (which was typically implemented using same context keys
anyway).

This hopefully makes the API simpler to use.

This is a breaking API change. RepoSpec and BaseURI now must be
provided as context values on each request to issuesapp handler.