Call loadTemplates in PostCommentHandler.
Fixes an issue where one couldn't react on newly posted comments
until the page was refreshed.

This was because state.IssueID is 0 in static template instead of
being the valid ID of the current issue, so the reactionsBar,
reactableID components had the wrong issue ID.

This bug was happened while working around the limitations of
templates, and won't happen in the future once the transition
to Go components for rendering HTML is complete.