Replace gorilla/mux with an app-specific mux.
This is a code refactor that does not change primary behavior but
removes the reliance on an external gorilla/mux dependency.

This codebase used gorilla/mux for its rounding needs instead of the
simpler http.ServeMux from the very beginning, because at that time,
I did not know another way of achieving routing with pattern matching,
and I needed it for routes such as:

	GET  /{id:[0-9]+}
	POST /{id:[0-9]+}/edit
	POST /{id:[0-9]+}/comment
	POST /{id:[0-9]+}/comment/{commentID:[0-9]+}

However, I did not enjoy using the API of gorilla/mux, because to me,
it is complicated and non-intuitive. I've since found a simpler way
to implement equivalent behavior by creating an app-specific mux, and
this change implements that.

Writing this requires a little more time initially, but the benefits
include more readable and predictable code.

There were specific behaviors I could not achieve with gorilla/mux, not
without spending a lot of time digging into its internals (because its
documentation did not answer it), quite likely needing to make changes
to it. But it becomes trivial to accomplish now, because the logic for
routing is completely self-contained and readable.

Additionally, the new code is more type safe. The URL route variables
no longer need to be looked up from a map. E.g., something like this:

	cr := issues.CommentRequest{
		ID:   uint64(mustAtoi(vars["commentID"])),
		Body: &body,
	}

Gets replaced by:

	cr := issues.CommentRequest{
		ID:   commentID,
		Body: &body,
	}

Because commentID is now a uint64 parameter to the handler. A missing
"commentID" would've been a runtime error previously, but now it'd be
a compile-time error.

Add an initial simple test for routes returning the expected HTTP status
code.

This change is related to a blog post I really enjoyed on this topic at
https://blog.merovius.de/2017/06/18/how-not-to-use-an-http-router.html
(/cc @Merovius).