Move assets/_data/script package to frontend; move assets/_data to _data.
Package script is a Go package that happens to be for frontend use,
meaning it's meant to run in a browser environment with DOM available.
However, it's still a proper Go package that can be built and tested
with the gc compiler, as well as checked by vet and other Go tools.
This has already paid off by helping find and fix a minor issue:

	$ go tool vet .
	frontend/scroll.go:73: no formatting directive in Errorf call

Move the package to a shorter, friendlier import path, and remove the
// build +js build constraint because it's not necessary. This way, it
doesn't get ignored by ./... import path pattern (previously, it was
ignored because it was in a _data directory, and because of js build
constraint).

Moving assets/_data to _data reduces nesting and makes it easier to
work on _data contents. Package assets just happens to use some _data
files.

This also improves consistency with other projects.

Regenerate.