Constrain nav style, remove global anchor style.
Constrain nav style so it doesn't interfere with potential other
elements on the page (e.g., in the user provided header).

Remove global anchor style because it's not used and not needed.
The app itself takes care of it individually where needed.

Related to above, move "text-decoration: none;" to apply to all nav
anchor modes, because without the global anchor style, this is now
needed. It should've been this way from the beginning, it was
overlooked due to global anchor style overriding it.

Regenerate.