component: Don't force opening links in new tab.
We're trying to optimize for best user experience rather than
retention. It should be simpler and friendlier for users to have the
links navigate to target right away, as that's the most common and
expected behavior of a link. If they wish, they can open link in new
tab themselves.